avro-generate-go: add tests for error conditions
This means we can write tests for the non-happy path.

Writing the first test uncovered an existing bug: because gogen-avro doesn't check all types when compiling (see https://github.com/actgardner/gogen-avro/issues/117), we can have analysis errors that aren't just as a result of internal errors, so we fix that issue too, by checking properly for type compatibility. This also meant that we could remove a hack to do with Set on a union.