support external verification of round trip tests
This allows us to verify the Go Avro behaviour against
the canonical Java implementation (and potentially other
implementations too).

We fix the tests that require fixing because of this.

The script I've been using for AVRO_VERIFY is this:

	#!/usr/bin/env rc

	if(! ~ $#* 3) {
		echo usage: avroverify in-schema out-schema in-data >[1=2]
		exit 1
	}
	inSchema=$1
	outSchema=$2
	inData=$3
	dir=`{mktemp -d}
	if (! echo $inData | avro fromjson --schema $inSchema - > $dir/data.out) {
		echo fromjson failed >[1=2]
		exit 1
	}
	if (! avro tojson --reader-schema $outSchema $dir/data.out) {
		echo tojson failed >[1=2]
		exit 1
	}
	rm -r $dir

where the `avro` command is itself a wrapper around the `avro-tools.jar` file:

	java \
		'-Dlog4j.configuration=file://'^<{echo 'log4j.rootLogger=off'} \
		--add-opens 'java.security.jgss/sun.security.krb5=ALL-UNNAMED' \
		-jar $AVRO_TOOLS_JAR $*

Note: the script above relies on functionality not yet available in `avro-tools` until https://github.com/apache/avro/pull/785 lands.