allow different names in non-union references
This allows us to read a record without using the same name
when reading it, which matches the Java implementation and
means we can go with our current plan for encoding CloudEvent
metadata inside Kafka messages.

Note: this is reliant on the relevant update to the forked version of gogen-avro.