rationalize TODOs
We remove TODOs that have been done and link most of the rest to
issues in the github repo. All outstanding issues
for the items in the TODO file now have github issues, so we can
remove that file.

We also fix the fact that comments weren't correctly being generated
on union fields by the code generator.