avro: cache type encoders at top level only
We'd like to provide the possibility to pair the encoding for a top level
type with a chosen schema.  That means we can't assume there's always
a specific association between Go type and Avro encoder for that type,
so we want to remove the use of the global cache in the body of the
encoder builder.

This is the first step towards allowing external definitions in generated
schema code.

This code doesn't change any externally-visible semantics.