cmd/avro-generate-go: support external types
We now support a metadata annotation: `go.package`
to indicate that instead of generating the code for a type,
a reference will be made to an externally defined Go type.

We also support a metadata annotation `go.name` to
indicate the name of the external Go type.

That type must encode to a compatible Avro schema, where
compatibility is defined more strictly than for Avro - the idea
is that a Go type already has stricter compatibility requirements,
we require that all struct fields must have the same name and type except that
struct fields may be added.

To find out information on the external Go type, we generate
and run a temporary Go program to introspect the types.
This is necessary because we can't find out which Go types
are enums without running their `String` method.