avro: fix program analysis
The current algorithm was too naive, and didn't cater for
the fact that the same PC could be reached via several
different call sites, so stopping the execution whenever
this happened caused some sites not to be reached.

Fix this by stopping executiion only when we
have a call stack that's that's the same.
To prevent infinite loops when we have a recursive
call, we also keep a record of what calls we've made
in the current stack and ignore a call if it's already
there.