avro: add CanonicalString method
Because of https://github.com/confluentinc/schema-registry/issues/1348, we
need to be sure that schemas added to the registry and searched in the registry
don't have extra metadata that can cause the search to fail.

To do that, we implement the canonical parsing normalization as specified
in the [Avro spec](https://avro.apache.org/docs/1.9.1/spec.html#Transforming+into+Parsing+Canonical+Form), with the addition that defaults can also be retained.