allow retaining logical types in CanonicalString
Use this functionality to retain logical types in the schemas
stored in the registry. This means that we'll be able to do
schema-resolution checks that logical types haven't inappropriately
changed (for example that the scale on a decimal value is the same)
or generate appropriate conversion code.

Somewhat gratuitously (but it's very early days, so hopefully OK)
we change the naming of the canonicalization constants to use
a `Retain` prefix instead of `Leave`, as that seems less ambiguous.