support empty Go struct types
Some Avro tooling doesn't allow empty records, despite not
being prohibited by the Avro spec, but it's useful to be able
to have them (for example to mark a particular kind of message
in a union that might not have any associated data).

So we allow empty Go structs and use a special form for them
in the Avro schemas (a single field with name `_` and type `"null"`).

We recognize that type both when encoding and decoding.