avroregistry: better error message on non-JSON response
When a service is unavailable or a proxy isn't working, we won't get the expected JSON response, so we were getting a confusing error message.

This change uses the [httprequest](https://pkg.go.dev/gopkg.in/httprequest.v1?tab=doc) package which does a much better job of trying to produce a good error message in the face of unexpected content.