cmd/avrogo: fix output file names
We fix an issue where a relative path passed as argument to the `-d`
flag would cause `avrogo` to report that the file doesn't exist,
and we also implement an algorithm to calculate unambiguous
output file names when passed several schema files with the same
base file name in almost all cases.

We also create the output directory if it doesn't already exist.