Update generate.go
The iteration order over maps is not specified and is not guaranteed to be the same from one iteration to the next. Therefore, iterating through the namespace.Definitions map may sometimes not work with the sort.Search() function. This tweak seems to be much simpler to understand and makes more sense for the purpose of the shouldImportAvroTypeGen() function.